changed directory names, made some modifications
