--- examples/app/bokeh-decision-tree/Decision_Tree/Plot/README.md
@@ -1,5 +1,0 @@
-# Bokeh Operations
-
-* instance.py - Singleton class for data set
-* get_data.py - Parse source, pass to tree files, get to plot data
-* plot_decision_tree.py- Plotting operations
