--- examples/app/bokeh-decision-tree/README.md
@@ -1,4 +1,0 @@
-# Decision Tree Visualizer
-Interactive Decision Tree using Bokeh
-# Usage
-```bokeh serve --show bokeh-decision-tree```
