+++ examples/app/decision_tree/README.md
@@ -1,0 +1,4 @@
+# Decision Tree Visualizer
+Interactive Decision Tree using Bokeh
+# Usage
+```bokeh serve --show bokeh-decision-tree```
